# Conflicts:
#	Diabetes/app/src/main/java/edu/uwm/android/diabetes/Activities/DietActivity.java
#	Diabetes/app/src/main/res/layout/activity_regimen.xml

+++ Diabetes/app/src/main/java/edu/uwm/android/diabetes/Activities/ExerciseActivity.java
@@ -67,18 +67,7 @@ public class ExerciseActivity extends AppCompatActivity {
         showSharedPreferences();
 
         //for auto complete
-        allExercise =new ArrayList<>();
-        Cursor cursor = databaseHandler.getData(new Exercise());
-        if(cursor.moveToFirst()){
-            do{
-                if(!allExercise.contains(cursor.getString(2)))
-                    allExercise.add(cursor.getString(2));
-
-            }while(cursor.moveToNext());
-        }
-        ArrayAdapter<String> adapter = new ArrayAdapter<String>
-                (this,android.R.layout.simple_list_item_1,allExercise);
-        exerciseDescription.setAdapter(adapter);
+        allExercise =autoCom();
 
 
         exerciseTime.setOnClickListener(new View.OnClickListener() {
@@ -109,9 +98,9 @@ public class ExerciseActivity extends AppCompatActivity {
         });
 
         if(getIntent().getIntExtra("exerciseId",-1) == -1){
-            updateExercise.setVisibility(View.INVISIBLE);
+            updateExercise.setEnabled(false);
         }else{
-            addExercise.setVisibility(View.INVISIBLE);
+            addExercise.setEnabled(false);
             exerciseDescription.setText(getIntent().getStringExtra("exerciseDescription"));
             String dateAndTime = getIntent().getStringExtra("exerciseDate");
             exerciseDate.setText(dateAndTime.substring(0,8));
@@ -122,21 +111,32 @@ public class ExerciseActivity extends AppCompatActivity {
 
             @Override
             public void onClick(View v) {
-                Exercise exercise = new Exercise();
-                exercise.setDescription(exerciseDescription.getText().toString());
-                exercise.setDate(exerciseDate.getText().toString() +" " + exerciseTime.getText().toString());
-                userName =  getIntent().getStringExtra("userName");
-                 databaseHandler.add(exercise,userName);
-                Toast.makeText(ExerciseActivity.this, "Description "+ exerciseDescription.getText().toString() + " Date "+
-                                exerciseDate.getText().toString()+" Added",
-                        Toast.LENGTH_LONG).show();
-                exerciseDate.getText().clear();
-                exerciseDescription.getText().clear();
-                SharedPreferences sp = getSharedPreferences("exerciseInfo", Context.MODE_PRIVATE);
-                SharedPreferences.Editor editor = sp.edit();
-                editor.clear();
-                editor.commit();
-
+                if(exerciseDescription.getText().toString().equals("")){
+                    Toast.makeText(ExerciseActivity.this, "Please add a description",
+                            Toast.LENGTH_LONG).show();
+                }else if(exerciseDate.getText().toString().equals("")){
+                    Toast.makeText(ExerciseActivity.this, "Please add a date",
+                            Toast.LENGTH_LONG).show();
+                }else if(exerciseTime.getText().toString().equals("")){
+                    Toast.makeText(ExerciseActivity.this, "Please add a time",
+                            Toast.LENGTH_LONG).show();
+                }else {
+                    Exercise exercise = new Exercise();
+                    exercise.setDescription(exerciseDescription.getText().toString());
+                    exercise.setDate(exerciseDate.getText().toString() + " " + exerciseTime.getText().toString());
+                    userName = getIntent().getStringExtra("userName");
+                    databaseHandler.add(exercise, userName);
+                    allExercise =autoCom();
+                    Toast.makeText(ExerciseActivity.this, "Description " + exerciseDescription.getText().toString() + " Date " +
+                                    exerciseDate.getText().toString() + " Added",
+                            Toast.LENGTH_LONG).show();
+                    exerciseDate.getText().clear();
+                    exerciseDescription.getText().clear();
+                    SharedPreferences sp = getSharedPreferences("exerciseInfo", Context.MODE_PRIVATE);
+                    SharedPreferences.Editor editor = sp.edit();
+                    editor.clear();
+                    editor.commit();
+                }
             }
         });
 
@@ -149,6 +149,7 @@ public class ExerciseActivity extends AppCompatActivity {
                 int id =getIntent().getIntExtra("exerciseId",-1);
                 if(id != -1) {
                     databaseHandler.update(getIntent().getIntExtra("exerciseId", -1), exercise,getIntent().getStringExtra("userName"));
+                    allExercise =autoCom();
                     getIntent().removeExtra("exerciseDate");
                     getIntent().removeExtra("exerciseDescription");
                     getIntent().removeExtra("exerciseId");
@@ -172,7 +173,6 @@ public class ExerciseActivity extends AppCompatActivity {
                 DateDialog();
             }
         });
-        cursor.close();
     }
 
     public void DateDialog() {
@@ -228,4 +228,22 @@ public class ExerciseActivity extends AppCompatActivity {
 
     @Override
     protected void onRestart() {super.onRestart();}
+
+    private ArrayList<String> autoCom(){
+        ArrayList<String> ret =new ArrayList<>();
+
+        Cursor cursor = databaseHandler.getData(new Exercise());
+        if(cursor.moveToFirst()){
+            do{
+                if(!ret.contains(cursor.getString(2)))
+                    ret.add(cursor.getString(2));
+
+            }while(cursor.moveToNext());
+        }
+        ArrayAdapter<String> adapter = new ArrayAdapter<String>
+                (this,android.R.layout.simple_list_item_1,ret);
+        exerciseDescription.setAdapter(adapter);
+        cursor.close();
+        return ret;
+    }
 }
