+++ Diabetes/app/src/main/java/edu/uwm/android/diabetes/Activities/MedicineActivity.java
@@ -13,15 +13,11 @@ import android.widget.AutoCompleteTextView;
 import android.widget.Button;
 import android.widget.DatePicker;
 import android.widget.EditText;
-import android.widget.ImageButton;
 import android.widget.TimePicker;
 import android.widget.Toast;
-
 import java.util.ArrayList;
 import java.util.Calendar;
-
 import edu.uwm.android.diabetes.Database.DatabaseHandler;
-import edu.uwm.android.diabetes.Database.Diet;
 import edu.uwm.android.diabetes.Database.Exercise;
 import edu.uwm.android.diabetes.Database.Medicine;
 import edu.uwm.android.diabetes.R;
@@ -64,18 +60,7 @@ public class MedicineActivity extends AppCompatActivity {
         showSharedPreferences();
 
         //for auto complete
-        allMedicine =new ArrayList<>();
-        Cursor cursor = databaseHandler.getData(new Medicine());
-        if(cursor.moveToFirst()){
-            do{
-                if(!allMedicine.contains(cursor.getString(2)))
-                    allMedicine.add(cursor.getString(2));
-
-            }while(cursor.moveToNext());
-        }
-        ArrayAdapter<String> adapter = new ArrayAdapter<String>
-                (this,android.R.layout.simple_list_item_1,allMedicine);
-        medicineDescription.setAdapter(adapter);
+        allMedicine =autoCom();
 
 
 
@@ -107,10 +92,10 @@ public class MedicineActivity extends AppCompatActivity {
         });
 
         if(getIntent().getIntExtra("medicineId",-1) == -1){
-            updateMedicine.setVisibility(View.INVISIBLE);
+            updateMedicine.setEnabled(false);
             System.out.println("Update is invisible");
         }else{
-            addMedicine.setVisibility(View.INVISIBLE);
+            addMedicine.setEnabled(false);
             medicineDescription.setText(getIntent().getStringExtra("medicineDescription"));
             String dateAndTime = getIntent().getStringExtra("medicineDate");
             medicineDate.setText(dateAndTime.substring(0,8));
@@ -121,21 +106,32 @@ public class MedicineActivity extends AppCompatActivity {
 
             @Override
             public void onClick(View v) {
-                Medicine medicine = new Medicine();
-                medicine.setDescription(medicineDescription.getText().toString());
-                medicine.setDate(medicineDate.getText().toString()+" "+ medicineTime.getText().toString());
-                userName =  getIntent().getStringExtra("userName");
-                databaseHandler.add(medicine, userName);
-                Toast.makeText(MedicineActivity.this, "Description "+ medicineDescription.getText().toString() + " Date "+
-                                medicineDate.getText().toString()+" Added",
-                        Toast.LENGTH_LONG).show();
-                medicineDate.getText().clear();
-                medicineDescription.getText().clear();
-                SharedPreferences sp = getSharedPreferences("medicineInfo", Context.MODE_PRIVATE);
-                SharedPreferences.Editor editor = sp.edit();
-                editor.clear();
-                editor.commit();
-
+                if(medicineDate.getText().toString().equals("")){
+                    Toast.makeText(MedicineActivity.this, "Please add a date",
+                            Toast.LENGTH_LONG).show();
+                }else if(medicineDescription.getText().toString().equals("")){
+                    Toast.makeText(MedicineActivity.this, "Please add a description",
+                            Toast.LENGTH_LONG).show();
+                }else if(medicineTime.getText().toString().equals("")){
+                    Toast.makeText(MedicineActivity.this, "Please add a time",
+                            Toast.LENGTH_LONG).show();
+                }else {
+                    Medicine medicine = new Medicine();
+                    medicine.setDescription(medicineDescription.getText().toString());
+                    medicine.setDate(medicineDate.getText().toString() + " " + medicineTime.getText().toString());
+                    userName = getIntent().getStringExtra("userName");
+                    databaseHandler.add(medicine, userName);
+                    allMedicine =autoCom();
+                    Toast.makeText(MedicineActivity.this, "Description " + medicineDescription.getText().toString() + " Date " +
+                                    medicineDate.getText().toString() + " Added",
+                            Toast.LENGTH_LONG).show();
+                    medicineDate.getText().clear();
+                    medicineDescription.getText().clear();
+                    SharedPreferences sp = getSharedPreferences("medicineInfo", Context.MODE_PRIVATE);
+                    SharedPreferences.Editor editor = sp.edit();
+                    editor.clear();
+                    editor.commit();
+                }
             }
         });
         updateMedicine.setOnClickListener(new View.OnClickListener() {
@@ -147,6 +143,7 @@ public class MedicineActivity extends AppCompatActivity {
                 int id =getIntent().getIntExtra("medicineId",-1);
                 if(id != -1) {
                     databaseHandler.update(getIntent().getIntExtra("medicineId", -1), medicine,getIntent().getStringExtra("userName"));
+                    allMedicine =autoCom();
                     getIntent().removeExtra("medicineDate");
                     getIntent().removeExtra("medicineDescription");
                     getIntent().removeExtra("medicineId");
@@ -168,7 +165,6 @@ public class MedicineActivity extends AppCompatActivity {
                 DateDialog();
             }
         });
-        cursor.close();
     }
 
     public void DateDialog() {
@@ -225,5 +221,23 @@ public class MedicineActivity extends AppCompatActivity {
     @Override
     protected void onRestart() {
         super.onRestart();
+    }
+
+    private ArrayList<String> autoCom(){
+        ArrayList<String> ret =new ArrayList<>();
+
+        Cursor cursor = databaseHandler.getData(new Medicine());
+        if(cursor.moveToFirst()){
+            do{
+                if(!ret.contains(cursor.getString(2)))
+                    ret.add(cursor.getString(2));
+
+            }while(cursor.moveToNext());
+        }
+        ArrayAdapter<String> adapter = new ArrayAdapter<String>
+                (this,android.R.layout.simple_list_item_1,ret);
+        medicineDescription.setAdapter(adapter);
+        cursor.close();
+        return ret;
     }
 }
